refactor(excel-checker): moves storage service in new file service
